Data validation and UI
In this code, I've added data validation and UI to the Portfolio class.
For example, in the addHolding method, I've added a check to make sure that the user has entered a symbol and a valid amount. If not, it shows an alert message.
I have added similar validation in other methods like removeHolding,setAlert,removeAlert for user inputs.
retrieveHoldings method can be used to retrieve the data from the local storage or a database.
getValue method can be used to calculate the total value of the portfolio.
checkAlerts method can be used to check alerts and notify the user.