News and alerts
In this code, the CCE class includes a function for subscribing to news and alerts related to specific cryptocurrencies. This function uses a combination of web scraping and API calls to retrieve the latest news and updates from various crypto news sources and social media platforms. The data is then parsed and stored in an object, which can be accessed and displayed by other functions within the CCE class. For example, a function could be created to display the latest news headlines related to a specific coin, or to send notifications to the user when there is significant news or price movement for a coin in their portfolio.