Supporting more exchanges
In this code, the Portfolio class has a exchanges property that is an object containing Exchange objects. The addExchange method is used to add new exchanges to the portfolio, and the removeExchange method can be used to remove them. The getExchange method allows for retrieving the specific exchange object. The Exchange class has getHoldings and placeOrder methods for retrieving the user's holdings and placing an order on the exchange. With this implementation the user can add multiple exchange and can get their holdings and place orders on them.